bot/requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
frozenlist==1.8.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.1
propcache==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.2.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.22.0
